Create PathfindingHelper class.
