+++ package.json
@@ -13,7 +13,7 @@
 	"source": "src/index.js",
 	"scripts": {
 		"dev": "concurrently \"microbundle watch --target browser\" \"serve --listen 5000\"",
-		"docs": "documentation build src/index.js --shallow -f md | replace-between --target README.md --token API",
+		"docs": "documentation build src/Pathfinding.js src/PathfindingHelper.js --shallow -f md | replace-between --target README.md --token API",
 		"dist": "microbundle --target browser",
 		"version": "npm run dist && && npm test && git add -A dist",
 		"postversion": "git push && git push --tags && npm publish",
